crawler/requirements.txt
requests
bs4
langdetect
psutil
snowflake-snowpark-python
pandas
python-dotenv